+++ apps/web/src/components/auction-edit/form-edit.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import CreateAuctionForm from '@/components/auction-form/form-create';
+import CreateAuctionForm from '@/components/auction-create/form-create';
 import useEditAuction from '@/hooks/auction/useEditAuction';
 import { Button } from '@/components/ui';
