+++ apps/web/src/components/common/auction-form/attach-images-thumbnail.tsx
