+++ apps/web/src/components/common/auction-form/category-select.tsx
