+++ apps/web/src/components/common/auction-form/error-msg.tsx
